+++ eslint.config.mjs
@@ -62,6 +62,7 @@ export default tseslint.config(
       'gsd-core/bin/lib/host-integration.cjs',
       'gsd-core/bin/lib/handshake-serialized.cjs',
       'gsd-core/bin/lib/host-integration-sdk.cjs',
+      'gsd-core/bin/lib/install-effort-resolver.cjs',
       'gsd-core/bin/lib/install-engine.cjs',
       'gsd-core/bin/lib/capability-loader.cjs',
       'gsd-core/bin/lib/capability-source.cjs',
